+++ snowddl/resolver/clone_table.py
@@ -1,4 +1,4 @@
-from snowddl.blueprint import TableBlueprint
+from snowddl.blueprint import TableBlueprint, SchemaObjectIdent
 from snowddl.resolver.abc_schema_object_resolver import AbstractResolver, ResolveResult, ObjectType
 
 
@@ -26,6 +26,7 @@ class CloneTableResolver(AbstractResolver):
 
     def get_databases_for_clone(self):
         databases_for_clone = {}
+        clone_source_env_prefix = self.engine.settings.clone_source_env_prefix
 
         cur = self.engine.execute_meta("SHOW DATABASES")
 
@@ -34,12 +35,24 @@ class CloneTableResolver(AbstractResolver):
             if r["origin"]:
                 continue
 
-            # Skip databases without destination for cloning
-            if f"{self.config.env_prefix}{r['name']}" not in self.engine.schema_cache.databases:
+            src_database = str(r["name"])
+
+            if clone_source_env_prefix:
+                # Skip everything which does not start with source prefix
+                if not src_database.startswith(clone_source_env_prefix):
+                    continue
+
+                dst_database = f"{self.config.env_prefix}{src_database.removeprefix(clone_source_env_prefix)}"
+            else:
+                dst_database = f"{self.config.env_prefix}{src_database}"
+
+            # Skip every source database without destination database for cloning
+            if dst_database not in self.engine.schema_cache.databases:
                 continue
 
-            databases_for_clone[r["name"]] = {
-                "database": r["name"],
+            databases_for_clone[dst_database] = {
+                "src_database": src_database,
+                "dst_database": dst_database,
             }
 
         return databases_for_clone
@@ -48,9 +61,9 @@ class CloneTableResolver(AbstractResolver):
         schemas_for_clone = {}
 
         cur = self.engine.execute_meta(
-            "SHOW SCHEMAS IN DATABASE {database:i}",
+            "SHOW SCHEMAS IN DATABASE {src_database:i}",
             {
-                "database": database["database"],
+                "src_database": database["src_database"],
             },
         )
 
@@ -59,12 +72,15 @@ class CloneTableResolver(AbstractResolver):
             if r["name"] == "INFORMATION_SCHEMA":
                 continue
 
-            # Skip schemas without destination for cloning
-            if f"{self.config.env_prefix}{r['database_name']}.{r['name']}" not in self.engine.schema_cache.schemas:
+            dst_schema = f"{database['dst_database']}.{r['name']}"
+
+            # Skip every source schema without destination schema for cloning
+            if dst_schema not in self.engine.schema_cache.schemas:
                 continue
 
-            schemas_for_clone[f"{r['database_name']}.{r['name']}"] = {
-                "database": r["database_name"],
+            schemas_for_clone[dst_schema] = {
+                "src_database": database["src_database"],
+                "dst_database": database["dst_database"],
                 "schema": r["name"],
             }
 
@@ -74,9 +90,9 @@ class CloneTableResolver(AbstractResolver):
         tables_for_clone = {}
 
         cur = self.engine.execute_meta(
-            "SHOW TABLES IN SCHEMA {database:i}.{schema:i}",
+            "SHOW TABLES IN SCHEMA {src_database:i}.{schema:i}",
             {
-                "database": schema["database"],
+                "src_database": schema["src_database"],
                 "schema": schema["schema"],
             },
         )
@@ -92,8 +108,9 @@ class CloneTableResolver(AbstractResolver):
             ):
                 continue
 
-            tables_for_clone[f"{self.config.env_prefix}{r['database_name']}.{r['schema_name']}.{r['name']}"] = {
-                "database": r["database_name"],
+            tables_for_clone[f"{schema['dst_database']}.{r['schema_name']}.{r['name']}"] = {
+                "src_database": schema["src_database"],
+                "dst_database": schema["dst_database"],
                 "schema": r["schema_name"],
                 "name": r["name"],
                 "is_transient": r["kind"] == "TRANSIENT",
@@ -119,20 +136,20 @@ class CloneTableResolver(AbstractResolver):
             query.append("TRANSIENT")
 
         query.append(
-            "TABLE IF NOT EXISTS {database_with_prefix:i}.{schema:i}.{table_name:i}",
+            "TABLE IF NOT EXISTS {dst_database:i}.{schema:i}.{name:i}",
             {
-                "database_with_prefix": f"{self.config.env_prefix}{row['database']}",
+                "dst_database": row["dst_database"],
                 "schema": row["schema"],
-                "table_name": row["name"],
+                "name": row["name"],
             },
         )
 
         query.append_nl(
-            "CLONE {database:i}.{schema:i}.{table_name:i}",
+            "CLONE {src_database:i}.{schema:i}.{name:i}",
             {
-                "database": row["database"],
+                "src_database": row["src_database"],
                 "schema": row["schema"],
-                "table_name": row["name"],
+                "name": row["name"],
             },
         )
 
@@ -140,6 +157,9 @@ class CloneTableResolver(AbstractResolver):
         r = cur.fetchone()
 
         if str(r["status"]).endswith("successfully created."):
+            self._drop_existing_policy_refs(
+                ObjectType.TABLE, SchemaObjectIdent("", row["dst_database"], row["schema"], row["name"])
+            )
             return ResolveResult.CREATE
 
         return ResolveResult.NOCHANGE
@@ -155,3 +175,59 @@ class CloneTableResolver(AbstractResolver):
             return True
 
         return False
+
+    def _drop_existing_policy_refs(self, object_type: ObjectType, object_name: SchemaObjectIdent):
+        cur = self.engine.execute_meta(
+            "SELECT * FROM TABLE(snowflake.information_schema.policy_references(ref_entity_domain => {object_type}, ref_entity_name => {object_name}))",
+            {
+                "object_type": object_type.singular_for_ref,
+                "object_name": object_name,
+            },
+        )
+
+        for r in cur:
+            if r["POLICY_KIND"] == "AGGREGATION_POLICY":
+                self.engine.execute_clone(
+                    "ALTER {object_type:r} {object_name:i} UNSET AGGREGATION POLICY",
+                    {
+                        "object_type": object_type.singular_for_ref,
+                        "object_name": object_name,
+                    },
+                )
+
+            elif r["POLICY_KIND"] == "MASKING_POLICY":
+                self.engine.execute_clone(
+                    "ALTER {object_type:r} {object_name:i} MODIFY COLUMN {column:i} UNSET MASKING POLICY",
+                    {
+                        "object_type": object_type.singular_for_ref,
+                        "object_name": object_name,
+                        "column": r["REF_COLUMN_NAME"],
+                    },
+                )
+
+            elif r["POLICY_KIND"] == "PROJECTION_POLICY":
+                self.engine.execute_clone(
+                    "ALTER {object_type:r} {object_name:i} MODIFY COLUMN {column:i} UNSET PROJECTION POLICY",
+                    {
+                        "object_type": object_type.singular_for_ref,
+                        "object_name": object_name,
+                        "column": r["REF_COLUMN_NAME"],
+                    },
+                )
+
+            elif r["POLICY_KIND"] == "ROW_ACCESS_POLICY":
+                policy_name = SchemaObjectIdent("", r["POLICY_DB"], r["POLICY_SCHEMA"], r["POLICY_NAME"])
+
+                self.engine.execute_clone(
+                    "ALTER {object_type:r} {object_name:i} DROP ROW ACCESS POLICY {policy_name:i}",
+                    {
+                        "object_type": object_type.singular_for_ref,
+                        "object_name": object_name,
+                        "policy_name": policy_name,
+                    },
+                )
+
+            else:
+                self.engine.logger.warning(
+                    f"Detected unknown policy type [{r['POLICY_KIND']} attached to cloned object [{object_name}]"
+                )